+++ utils.go
@@ -5,14 +5,14 @@ import (
 	"time"
 )
 
-func getTimeNow() int {
+func getTimeNow() float64 {
 	location, err := time.LoadLocation("Africa/Lagos")
 	if err != nil {
 		// Handle error
 		fmt.Println("Error loading location:", err)
-		return 0
+		return 0.0
 	}
 
 	now := time.Now().In(location)
-	return int(now.Unix())
+	return float64(now.Unix())
 }
